[#24]refactor: UpdateInfo 분리
