--- client/src/lib/api/types/api.ts
@@ -1,5 +1,0 @@
-interface ApiResponse {
-  statusCode: number;
-}
-
-export default ApiResponse;
