+++ client/src/lib/api/types/common.ts
@@ -1,0 +1,9 @@
+export interface ApiResponse {
+  statusCode: number;
+}
+
+export interface UpdateInfo {
+  idx: number;
+  createdAt: Date | string;
+  updatedAt: Date | string;
+}
