+++ client/src/lib/api/types/index.ts
@@ -1,4 +1,4 @@
-export { default as ApiResponse } from './api';
+export * from './common';
 export * from './error';
 export * from './auth';
 export * from './users';
